comments, credits, and randomness
Added comments to everything so that it is easier to follow, including a credits section at the top. Also now using srand(time(0)) to generate a random seed based on Unit timestamps. This fixes the issue with the randomness that was in the "known issues" section of the readme.md. Removed that part.

PS: also changes some things to say team 3 instead of team 4 bc for some reason I thought we were team 4.